requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
pymongo==4.5.0
